Remove the need for extra connections between agent and manager.
Replaces the forwarding of intercepted connections from agent to the
manager with a ConnTunnel so that no extra connections have to be made
to ports that aren't exposed in the manager's deployment manifest.

Also ensures that all outbound traffic from the client to the
traffic-manager is routed to intercepted agents so that those requests
originates from a pod in the currently intercepted namespace.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>