Add an agent injecting mutating webhook to the traffic manager.
This commit adds a mutating webhook that injects a traffic-agent to
pods that have the following annotation:

    telepresence.getambassador.io/traffic-agent: enabled

The installer will henceforth ignore deployments that have this
annotation and consider them to already have an agent installed.

The webhook is using `kates.List()` to list Service objects which
means that the cluster must be configured to allow the traffic-manager
to do that. A sample `k8s/agent-injector-rbac.yaml` is included that
provides the needed credentials.

A lot of code used by the agent installer is reused in the injector
and was therefore moved into the `install` package that they
both now share.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>